Reorganize namespaces so we can provide serialization automatically
I was recently convinced that providing a `data_readers.clj` in a lib
isn't a bad thing. We still had the problem of making sure that the
graph6 namespaces gets loaded though, and it couldn't be required from
the main graphs ns without a cyclic dependency. So I pulled in
potemkin so I could do an import-vars, and rearranged the namespaces
appropriately.

Now serialization should just work.